Added Database to new-order-form
I added a dropbox in new_order_form.php. HOWEVER, there is an account
error when referring to a manager or sales agent. The drop box lists
products and their prices. I didn't want to mess with style or acct
error, because that seemed like something someone else would like to
implement.